agent uči do određenog dijela i onda se vrti oko tog rezultata - napredak bio vidljiv do gen 15 od tada se prosjek bodova cvrti od 3.5-4.5 prvi put na 59 gen prešlo 5.0 iznad 70 gen se prosjek digao 4.5-5.5 i čak prešao 6.0  - iako najbolji se vrti oko 20 stalno

+++ Zmijica/src/GP.cpp
@@ -262,10 +262,10 @@ float GP::evaluateTree(GPNode* root) {
         return (float)l_val / (float)r_val;
     
     if (root->getKey_value() == 4)
-        return std::min((float)l_val, (float)r_val);
+        return (float)(std::min((float)l_val, (float)r_val));
 
     if (root->getKey_value() == 5)
-        return std::max((float)l_val, (float)r_val);
+        return (float)(std::max((float)l_val, (float)r_val));
 }
 
 int GP::getMaxDepth() {
